+++ unit/BalancingWalker.h
@@ -51,7 +51,7 @@ private:
     S32 turn;
     bool standControlMode;
     int offset;
-    double offsetRightEncBeforeInit;
+    long offsetRightEncBeforeInit;
 
     S32 rightWheelEncOffset;	// 倒立制御復帰時にエンコーダをリセットしないように実装するとき必要 今は使ってない
     S32 leftWheelEncOffset;		// 倒立制御復帰時にエンコーダをリセットしないように実装するとき必要 今は使ってない
